Add `facts` and `user_agent` properties to the node
These provide a mechanism to access the facts, and the user agent, data for
the current node.  These ensure that required configuration data is gathered
in a central location, rather than open-coding it in each consumer location.

For the moment the facts are simply the Facter facts, directly reported; we
have no custom configuration or exclusion.  In future, if we require those
changes this ensures there is one place, and only one place, to change that.

The user agent support gathers various version numbers to form the HTTP user
agent header -- and to allow the server to have a reasonable view of the state
of clients that are communicating with it.  As more tools or packages that
Razor depends on are added to the system, this can be extended to continue to
report the relevant information to allow the server to make decisions on next
required actions.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>