Implement the `register` command, and supporting code
This implements the `register` command, to submit data about the node to the
Razor server, and to handle the response which includes the next operation.

Presently, all functionality is present and correct other than the response
handling -- currently, stubbed out as the infrastructure to process commands
is not yet present.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>