Fix `bin/mk` error when command is not supplied
Using `to_s` on the value in ARGV would result, sadly, in the nil value
turning into an empty string -- and while it still failed, it did so with a
confusing report about a missing "empty" command.

This fixes that, and adds a test to ensure that the error reports are correct.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>