Implement server command execution using external binaries
This implements a simple pattern for execution of commands from the server:
a "whitelist" in the form of a directory on disk is used.  Any command will be
executed as requested by the server, provided that it exists in
that directory.

This makes it easy to both expose an entire command (eg: `reboot`, or `yum`),
or to provide a wrapper that limits the possible arguments to the command.

Nothing on the normal execution path is trusted, although that can be
overridden by setting, eg, `razor.commands=/bin` in the configuration.

Right now only a single directory is supported, although it would be
reasonably easy to extend this to a full set of paths, or provide other
mechanisms to extend the whitelist for other commands.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>