Fix lookup of config values from kernel command line and config file
Both of these would always find a match as long as there was at least one
config setting in the respective source because String.casecmp returns
always a non-nil value. Check properly that strings a and b match case
insensitively by checking for "a.casecmp(b) == 0" instead of
"a.casecmp(b)".