Comprehensive network interface support: DHCP on all the things!
This implements DHCP configuration for all network interfaces.  Literally all
of them, in fact: anything other than the loopback interface should
automatically be:

A) added to `/etc/sysconfig/network-scripts/ifcfg-*`.
B) configured for DHCP.
C) configured not to fail if a DHCP offer is not received, but to
   keep trying until one finally arrives on that interface.
D) brought up with the `network.service` systemd unit.

This is fully automatic -- no user configuration is supported, and it is
fully driven out of udev device discovery.

This also means that it should work with, eg, hotplug network interfaces, USB
devices that show up late, layered or lazy network interfaces (if anyone *ever*
implements one of them, though VPN seems ... possible, if crazy), etc.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>