Use UTC timezone for the Microkernel
The Microkernel was previously configured in the US/Eastern timezone -- which
is not the most helpful source of timestamps on the west coast, let alone
elsewhere in the world.

This updates the image to use UTC, which is probably the least worst portable
timestamp / system clock timezone option available to us.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>