Cut ~ 98MB out of the pregenerated locale archive
Instead of unpacked locale data, Fedora 19 uses a binary archive that is
memory mapped into every running process.  Conveniently, this avoids a lot of
the overheads of the older, separate files.

Inconveniently for us, they build it with all the locales in, and that weighs
in at about 100MB of data.  Now, we purge everything but en_US locale data,
giving us a resultant image of 2.1MB.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>