Remove more packages, more cleanly
A number of packages are only needed during the build process; this removes a
much broader set and shaves some more MB out of the final image.

It also moves the package removal *before* we destroy their unpacked locale
data, removing a stream of complaints from the packaging tools -- and making
it possible to see if something actually went wrong!

Signed-off-by: Daniel Pittman <daniel@rimspace.net>