Truncate logfiles before building the image
A number of logfiles have data -- a piddly amount -- from during the build
process, not the current run of the Microkernel.

While the data savings are trivial, this ensures that what you get in the logs
reflects what *actually* happened on this machine, without polluting it based
on whatever random system the Microkernel ISO was built on.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>