Remove /boot from inside the Microkernel images
The build process for the microkernel extracts the kernel and other components
out of the target image, and writes them to the ISO9660 filesystem.

Once booted, the running in-memory system contained another copy of the same
kernel and support code -- useful for installing to disk, but absolutely
worthless for anything at runtime.

This commit deletes those files before the final archive is built; they are
already copied into the "outside" ISO9660 filesystem by this point.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>