Disable SSH daemon by default
This disables starting the SSH daemon in the Microkernel instance by default,
which seems like the best security strategy.  We can enable a Microkernel
command in future to start it for us, or perhaps offer some kernel
command-line feature to enable it.

This also destroys -- if present -- any SSH host keys in the final image.
The startup process for sshd will generate new, unique microkernel keys.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>