Use NetworkManager to handle network device configuration
In order to have network connections managed effectively, we use
NetworkManager.  This is ~ 5.9MB of space in the image, but it also means that
we are (A) using the recommended and default upstream configuration, and (B)
no longer responsible for doing all the network management ourselves.
This is, overall, a big win for everyone.

Also, this opens the door to allowing for more complex configurations such as
802.1x secured network links, VPN connectivity for communication with the
host, and so forth -- should we decide we need it.

Ultimately, though, that as the upstream project write:

   "Fedora now by default relies on NetworkManager for network
    configuration. This is the case also for minimal installations and server
    installations. We are trying to make NetworkManager as suitable for this
    task as possible."

I hope that doesn't offend.  Dropping this in just works!

Signed-off-by: Daniel Pittman <daniel@rimspace.net>