Add unzip(1) to the MK, used to unpack the runtime extension file
We serve a zip file from the server in order to allow runtime extension of the
Microkernel runtime; the most efficient mechanism to allow unpacking that is
the unzip(1) binary, which also has happily consistent feature support and is
maintained upstream in the distro.

This adds the binary to the ISO image, nothing else.

https://tickets.puppetlabs.com/browse/RAZOR-42
Signed-off-by: Daniel Pittman <daniel@rimspace.net>