Add mk-update wrapper to LiveCD
This adds an mk-update wrapper script to the LiveCD which is designed to
download, and unpack if changed, the extension zip file.  It will then execute
the original MK application with the appropriate environment variables set.

Separating this out into a separate executable helps ensure that we can always
run the update process, even if it drops something onto disk that would cause
the mk executable to fail.

This makes it significantly easier to recover from a failure than if they were
bundled together and this feature could be damaged or disabled by the change.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>