Add mk config command to extract a configuration variable
This makes it possible for external tools to access the configuration
discovery and parsing functionality embedded in the MK Ruby script, without
having to be written in Ruby.

This is used by the extension system to allow the external shell script to
obtain this data without reimplementing the configuration parsing externally.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>